+++ laws/src/main/scala/cats/tagless/laws/discipline/SemigroupalKTests.scala
